[refactor] :recycle: rename responseHelper yo requestHelper (#276)
