+++ src/controllers/nftController.ts
@@ -5,7 +5,7 @@ import { FastifyReply, FastifyRequest } from 'fastify';
 import { IUser } from '../models/user';
 import { Logger } from '../helpers/loggerHelper';
 import { isValidUrl } from '../helpers/validationHelper';
-import { getDynamicGas } from '../helpers/dynamicGasHelper';
+import { getDynamicGas } from '../helpers/paymasterHelper';
 import NFTModel, { INFT, INFTMetadata } from '../models/nft';
 import { getNetworkConfig } from '../services/networkService';
 import { createUserWithWallet } from '../services/userService';
@@ -13,7 +13,7 @@ import { getWalletByPhoneNumber } from '../services/walletService';
 import { sendMintNotification } from '../services/notificationService';
 import { SIGNING_KEY, defaultNftImage, DEFAULT_CHAIN_ID } from '../config/constants';
 import { uploadToICP, uploadToIpfs, downloadAndProcessImage } from '../services/uploadService';
-import { returnErrorResponse, returnSuccessResponse } from '../helpers/responseFormatterHelper';
+import { returnErrorResponse, returnSuccessResponse } from '../helpers/requestHelper';
 
 export interface NFTInfo {
   description: string;
