+++ src/controllers/tokenController.ts
@@ -4,7 +4,7 @@ import { FastifyReply, FastifyRequest, FastifyInstance } from 'fastify';
 import Token, { IToken } from '../models/token';
 import { Logger } from '../helpers/loggerHelper';
 import { BUN_ENV, SIGNING_KEY } from '../config/constants';
-import { returnErrorResponse, returnSuccessResponse } from '../helpers/responseFormatterHelper';
+import { returnErrorResponse, returnSuccessResponse } from '../helpers/requestHelper';
 
 /**
  * Creates a new token
